Fix landing page handlers for Next.js client props
